Compute the module docs for modules when it's missing
The problem was related to caching: By re-using the cache from the
file-system, we were also skipping the computation of the
Elm.Docs.Module of the module, which is needed to compute the module
name lookup table.

That created bugs like https://github.com/jfmengels/elm-review-unused/issues/75
where when a file is changed (but not its imports) then we'd compute the
lookup table again, but without the Elm.Docs.Module-s from the imports,
it would be incorrect and led to bad results.