Fix test around JSON decoding error message
Between Node.js v18 and v20, the error message when failing to decode JSON has been changed to be more helpful. This failed the test suite because the error message was still made for v18 environments.

Henceforth tests for this project should be run on v20+ versions of
Node.js.